### 2025-03-10 #7888 @editorialbot
[PRE REVIEW]: SounderPy: An atmospheric sounding visualization and analysis tool for Python
**Submitting author:** @kylejgillett (<a href="http://orcid.org/0009-0001-5528-5037">Kyle Gillett</a>)
**Repository:** https://github.com/kylejgillett/sounderpy
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.8
