### 2025-04-03 #7888 @kthyng
@pdebuyl Thank you! I just wrapped up my initial high level checks with this prompting and this is ready to hand off.
