+++ src/BUFKIT-STATIONS-MASTER.txt
@@ -296,7 +296,7 @@ WMO numbers must be sequential.      Coordinates are in normal degrees.
 121,259,45.35,-117.23,JOES,JOSEPH                          ,OR,1250,NWS W Reg 2-06
 122,260,41.59,-95.34,HNR ,HARLAN MUNICIPAL AIRPT          ,IA,375,NWS OKX 2-06
 123,261,43.78,-87.85,KSBM,SHEBOYGAN (ASOS)                ,WI,233,NWS MKE 4-06
-124,262,30.9,-102.85,KFST,FORT STOCKTON (ASOS)            ,TX,917,TAF 7-07
+124,262,30.9,-102.85,KSTK,FORT STOCKTON (ASOS)            ,TX,917,TAF 7-07
 125,263,33.64,-91.75,LLQ ,MONTICELLO ARPT (ASOS)          ,AR,82,TAF 7-7
 126,264,36.22,-92.28,BPK ,MOUNTAIN HOME                   ,AR,283,TAF 7-7
 127,265,31.38,-103.51,KPEQ,PECOS MUNICIPAL                 ,TX,796,TAF 7-7
