### 2025-05-12 #8087 @timothyas
Hi everyone, I have gone through my review checklist and I think the package, paper, and documentation being presented here are in really good shape. Installation was dead simple. The package implements a very clean API, the documentation is thorough, and the plots that can be produced present *a lot* of information in a highly condensed format. This is really impressive work @kylejgillett!! 👏

The only box that I felt like I was not able to fully check was the "community guidelines one". In the sounderpy documentation, it is clearly stated:

> Have an issue? You can [open a GitHub Issue](https://github.com/kylejgillett/sounderpy/issues) or just shoot me a DM on [Twitter](https://twitter.com/wxkylegillett)!

However, there are no guidelines for contributing to the package. Based on the JOSS guidelines I think this needs to be added. Along with that I would softly suggest putting the "Have an issue" statement in it's own section of the documentation (e.g. "Community" or "Reaching Out"). Right now it's in the "About the author" section, but I think having its own section in the table of contents would be a nice touch for the lost user 🙂.

Otherwise, the only suggestion I have is that I think it would be nice if the package was able to readily store data in netcdf format since this is such a popular file format for weather and climate data. Also, since xarray is one of the dependencies, it should be pretty straightforward to do so. However, let me just make clear that this is just a suggestion and I definitely don't view it to be required for JOSS acceptance. Especially since reading a csv to xarray is simple (use pandas, convert to xarray).

Again, great work @kylegillet! Let me know when you've added the contributing guidelines (or if they are there and I missed them, let me know where they are) and I'll check the last box and sign off on the review.
