### 2025-05-14 cb7ca47
add init files for v3.0.9; update docs to include contributing guidelines
