+++ src/sounderpy/rs_reflectivity.txt
@@ -1,0 +1,257 @@
+Product: BR
+Units:   DBZ
+Step:    5
+SolidColor:	-32.0		115	77	172
+SolidColor:	-31.5		115	78	168
+SolidColor:	-31.0		115	79	165
+SolidColor:	-30.5		115	81	162
+SolidColor:	-30.0		116	82	158
+SolidColor:	-29.5		116	84	155
+SolidColor:	-29.0		116	85	152
+SolidColor:	-28.5		117	86	148
+SolidColor:	-28.0		117	88	145
+SolidColor:	-27.5		117	89	142
+SolidColor:	-27.0		118	91	138
+SolidColor:	-26.5		118	92	135
+SolidColor:	-26.0		118	94	132
+SolidColor:	-25.5		119	95	128
+SolidColor:	-25.0		119	96	125
+SolidColor:	-24.5		119	98	122
+SolidColor:	-24.0		120	99	118
+SolidColor:	-23.5		120	101	115
+SolidColor:	-23.0		120	102	112
+SolidColor:	-22.5		121	103	108
+SolidColor:	-22.0		121	105	105
+SolidColor:	-21.5		121	106	102
+SolidColor:	-21.0		122	108	98
+SolidColor:	-20.5		122	109	95
+SolidColor:	-20.0		122	111	92
+SolidColor:	-19.5		123	112	88
+SolidColor:	-19.0		123	113	85
+SolidColor:	-18.5		123	115	82
+SolidColor:	-18.0		124	116	78
+SolidColor:	-17.5		124	118	75
+SolidColor:	-17.0		124	119	72
+SolidColor:	-16.5		125	121	69
+SolidColor:	-16.0		127	123	72
+SolidColor:	-15.5		129	125	75
+SolidColor:	-15.0		131	127	79
+SolidColor:	-14.5		133	130	82
+SolidColor:	-14.0		135	132	85
+SolidColor:	-13.5		137	134	89
+SolidColor:	-13.0		139	137	92
+SolidColor:	-12.5		141	139	96
+SolidColor:	-12.0		144	141	99
+SolidColor:	-11.5		146	144	102
+SolidColor:	-11.0		148	146	106
+SolidColor:	-10.5		150	148	109
+SolidColor:	-10.0		152	151	113
+SolidColor:	-9.5		154	153	116
+SolidColor:	-9.0		156	155	119
+SolidColor:	-8.5		158	158	123
+SolidColor:	-8.0		161	160	126
+SolidColor:	-7.5		163	162	130
+SolidColor:	-7.0		165	165	133
+SolidColor:	-6.5		167	167	136
+SolidColor:	-6.0		169	169	140
+SolidColor:	-5.5		171	172	143
+SolidColor:	-5.0		173	174	147
+SolidColor:	-4.5		175	176	150
+SolidColor:	-4.0		178	179	154
+SolidColor:	-3.5		173	175	153
+SolidColor:	-3.0		168	171	152
+SolidColor:	-2.5		163	167	151
+SolidColor:	-2.0		158	163	150
+SolidColor:	-1.5		154	159	149
+SolidColor:	-1.0		149	155	148
+SolidColor:	-0.5		144	151	147
+SolidColor:	 0.0		139	147	146
+SolidColor:	 0.5		135	144	145
+SolidColor:	 1.0		130	140	144
+SolidColor:	 1.5		125	136	143
+SolidColor:	 2.0		120	132	142
+SolidColor:	 2.5		115	128	142
+SolidColor:	 3.0		111	124	141
+SolidColor:	 3.5		106	120	140
+SolidColor:	 4.0		101	116	139
+SolidColor:	 4.5		96	112	138
[…]
