### 2025-05-14 #8087 @timothyas
Awesome, thanks @kylejgillett for adding that. With that addition I think the package is good to go, and I think the submission should be accepted to JOSS.

Once again, great work on this excellent package. I think it will be (and probably already is) very useful to the community! 👏
