### 2025-05-17 #8087 @trontrytel
Hi Everyone!

This is a great package and really awesome work on simplifying downloading the data and plotting. The package is really easy to use, and I'm sure it's very empowering for users to be able to quickly visualize the data they are working with. I confirm I was able to download and plot the data as described in the readme, and I'm neither a Python user, nor do I work with atmospheric sounding data.

I only wanted to ask about the automated tests, since I didn't see any CI setup in the repo. I saw the collab tutorials, which are great for new users and also allow to manually check the package functionality. Thats why I checked off the `Automated tests` box in the review checklist. An idea for the future could be to add some automated tests. A very simple (and maybe stupid) way would be to save some sort of reference plot(s) to a file and then diff the newly generated plots against that reference. This comment by no means should block the JOSS acceptance.
