### 2025-06-24 #8087 @pdebuyl
@kylejgillett sorry for the delay. I have started testing SounderPy and getting the software installed and following the first steps of the documentation works really well!

Then, I had to reflect on the inclusion of a copyright notice to your name in the figures output by SounderPy. While the software is indeed copyrighted to you, the copyright notice in the figures gives the impression that the figures generated by SounderPy also fall under your copyright, which is not the case as this is free software. Is this notice intended with this implied message?
