### 2025-07-24 984ba02
add egg.info, pycache to gitignore
