+++ src/sounderpy/plot.py
@@ -28,18 +28,26 @@ from PIL import Image
 import cartopy.crs as ccrs
 import cartopy.feature as cfeature
 import matplotlib.pyplot as plt
+from cartopy.io import img_tiles
 
-# DATA RETRIEVAL
-from xarray.backends import NetCDF4DataStore
-from xarray import open_dataset
-from siphon.catalog import TDSCatalog
-from netCDF4 import Dataset
-from matplotlib.colors import LinearSegmentedColormap
 
 # SOUNDERPY MODULES
 from .calc import sounding_params, vad_params
 from .utils import modify_surface, find_nearest, mag, mag_round
+from .radar_utils import *
 
+
+import os
+from contextlib import contextmanager, redirect_stderr, redirect_stdout
+@contextmanager
+def suppress_stdout_stderr():
+    """A context manager that redirects stdout and stderr to devnull"""
+    with open(os.devnull, 'w') as fnull:
+        with redirect_stderr(fnull) as err, redirect_stdout(fnull) as out:
+            yield (err, out)
+
+with suppress_stdout_stderr():
+    import pyart
 
 
 
@@ -56,115 +64,48 @@ from .utils import modify_surface, find_nearest, mag, mag_round
 """
 
 
+#########################################################################
+############################ PLOT UTILS #################################
+#########################################################################
+
+#########################################################################
+########################## HODO BOUNDARY ################################
+def plot_boundary(ax, angle_deg, color):
+    # line params
+    center_u = 0
+    center_v = 0
+    angle_rad = np.deg2rad(90-angle_deg)
+    # compute u & v deltas
+    delta_u = 200 * np.cos(angle_rad)
+    delta_v = 200 * np.sin(angle_rad)
+    # compute "u & v" values for the boundary
+    boundary_u = np.array([center_u - delta_u, center_u + delta_u])
+    boundary_v = np.array([center_v - delta_v, center_v + delta_v])
+    # plot the boundary
+    ax.plot(boundary_u, boundary_v, color=color, linestyle='--', linewidth=2, alpha=0.3, label='Boundary Axis')
+
+
+
+
+
+#########################################################################
+############################## FIGURES ##################################
+#########################################################################
+
+
+
 
 #########################################################################
 ########################## FULL SOUNDING ################################
 
 def __full_sounding(clean_data, color_blind, dark_mode, storm_motion, special_parcels=None, 
-                    show_radar=True, radar_time='sounding', map_zoom=2, modify_sfc=None,
-                    show_theta=False):
+                    radar='mosaic', radar_time='sounding', map_zoom=2, modify_sfc=None,
+                    show_theta=False, hodo_boundary=None):
     
     
     # record process time 
-    st = time.time()  
+    st = time.time()
 
-
-    
-    
-    # get radar mosaic data for map inset
-    def get_radar_mosaic(clean_data, map_zoom, radar_time):
-        '''
-        GET RADAR MOSAIC DATA FOR SOUNDING PLOT
-        
-        - gets data from NCEI for 'sounding time' or now
-        - data only goes back 1 month from current time
-        '''
-    
-        # search the NCEI database for the file closest to 
-        # the requested radar date/time
-        def find_file(target_time, file_names):
-            closest_index = None
-            closest_difference = float('inf')
-
-            for idx, file_name in enumerate(file_names):
[…]
