### 2025-08-20 #8087 @kylejgillett
Do you have a recommendation on how to mention the vendor-ing of Sharppy?
