### 2025-08-20 #8087 @pdebuyl
Replace "LICENSE" by "LICENSES" in which you explain that SounderPy is licensed under the MIT (+ text) and includes SHARPPY that is licensed under the BSD 3-Clause (+ mention the location of that license file). Alternatively, the MIT text can go in `LICENSE-MIT`, the BSD in `LICENSE-BSD` (both at the top-level of the repository, to avoid automatic tool from dismissing the BSD one), and an explanation in LICENSES (or in README, provided it is clear enough).
