### 2025-08-21 8d7d4e2
clean up repo for JOSS paper, add updates for v3.0.9
