### 2025-08-21 dc719a4
remove __pycache__, .egg, docs/build, docs/source from git history
