+++ src/sounderpy.egg-info/requires.txt
@@ -1,12 +1,0 @@
-cartopy>=0.21.0
-cdsapi>=0.7.3
-ecape-parcel>=1.2.0.2
-matplotlib>=3.3.0
-metpy>=1.5.1
-netcdf4>=1.6.4
-numpy>=1.20.0
-pandas>=1.2.0
-arm-pyart>=2.0.0
-siphon>=0.9
-scipy>=1.10.1
-xarray>=0.18.0
