+++ docs/build/html/_sources/examplescripts.rst.txt
@@ -1,24 +1,0 @@
-📝 Usage Examples
-==================
-
-Examples of SounderPy API in action!
-
-
-.. nblinkgallery::
-    :caption: Data Retrieval Examples
-    :name: rst-data-gallery
-
-    raob_data_example
-    reanalysis_data_example
-    acars_data_example
-    bufkit_data_example
-    custom_data_example
-
-
-.. nblinkgallery::
-    :caption: Plotting Examples
-    :name: rst-plot-gallery
-
-    sounding_example
-    hodograph_example
-    composite_example   
