+++ docs/source/helpertools.rst
@@ -1,287 +1,0 @@
-🛠️ Helper Tools
-================
-
-A collection of helper tools included within SounderPy.
-
-*************************
-
-
-Returning a dictionary of profile parameters
---------------------------------------------
-
-Return a dictionary of common calculated sounding paramters and special variables for a given vertical profile.
-
-Data returned in this dictionary include...
-      - an interpolated version of the profile
-      - basic parameters such as mixing ratio, theta-e, wet-bulb, etc
-      - Thermodynamic parameters such as SB, MU, ML parcel properties (LCL, LFC, EL, CAPE, CIN, etc), Entrainment CAPE, etc
-      - Kinematic parameters such as storm motion, storm-relative wind, streamwise vorticity, SRH, composite parameters, etc
-
-.. class:: sounding_params(clean_data, storm_motion='right_moving', modify_sfc=None)
-
-   :param clean_data: the dictionary of profile data to calculate profile parameters for (see :doc:`gettingdata`)
-   :type clean_data: dict, required
-   :param storm_motion: the storm motion used for calculations. Default is 'right_moving'. Custom storm motions are accepted as a `list` of `floats` representing direction and speed. Ex: ``[270.0, 25.0]`` where '270.0' is the *direction in degrees* and '25.0' is the *speed in kts*. See the :ref:`storm_motions` section for more details.
-   :type storm_motion: str or list of floats, optional
-   :param modify_sfc: a `dict` in the format ``{'T': 25, 'Td': 21, 'ws': 20, 'wd': 270}`` to modify the surface values of the ``clean_data`` dict.
-   :type modify_sfc: None or dict, optional, default is None
-
-
-   .. py:function:: .calc()
-
-      :return: special sounding parameters
-      :rtype: dict
-
-Example:
-
-.. code-block:: python
-
-   spy.sounding_params(clean_data).calc()
-
-.. note::
-   These calculated parameters are the same used on SounderPy plots. This function simply returns the master dictionary of all of these values.
-
-
-****************************************************************
-
-
-
-
-Saving data to a file
-----------------------
-
-.. py:function:: spy.to_file(file_type, clean_data, filename=None, convert_to_AGL=True)
-
-
-   Create a file of 'cleaned' SounderPy data
-
-   :param file_type: a `str` representing the file type you'd like to export data to.
-   :type file_type: str, required
-   :param clean_data: 'cleaned' SounderPy data `dict`
-   :type clean_data: dict, required
-   :param filename: the name you'd like to give the file
-   :type filename: str, required
-   :param convert_to_AGL: whether or not to convert height values to "above ground level" when saving to a file. Useful for CM1.
-   :type convert_to_AGL: bool, optional, default is `True`
-   :return: a file of SounderPy data.
-
-Example:
-	* File options include `csv`, `cm1`, & `sharppy`
-
-.. code-block:: python
-
-   spy.to_file('csv', clean_data)
-   spy.to_file('cm1', clean_data)
-   spy.to_file('sharppy', clean_data)
-
-
-***************************************************************
-
-
-Merging two profiles via weighted averaging
---------------------------------------------
-
-Merge two profiles together using weighted averaging of values along a common height axis to return a single profile.
-
-.. py:function:: spy.merge_profiles(profile_a, profile_b, weight_a)
-
-   :param profile_a: a SounderPy `clean_data`dictionary containing variable arrays for profile_a.
-   :type profile_a: dict, required
-   :param profile_b: a SounderPy `clean_data` dictionary containing variable arrays for profile_b.
-   :type profile_b: dict, required
[…]

[…]
