### 2025-08-22 4216547
edit LICENSES wording, spell-check paper.md, add test figure to docs
