+++ docs/_sources/about.rst.txt
@@ -111,6 +111,7 @@ Authors and Contributors
 	*CONTRIBUTOR: Scott Thomas, NWS Grand Rapids | VWP Hodograph, Buoy-sites listing*
 	*CONTRIBUTOR: Amelia R H Urquhart, University of Oklahoma | ecape-parcels library*
 	*CONTRIBUTOR: Daryl Herzmann, Iowa State University | SounderPy Feedstock for conda-forge*
+	*CONTRIBUTOR: Ryan Vandersmith | Stepwise CAPE/CIN plot*
 
 ***************************************************************
 
@@ -122,7 +123,7 @@ Citing SounderPy
 
 
 	in AMS format:
-	     Gillett, K., 2024: SounderPy: Vertical Profile Data Retrieval & Analysis Tool for Python (Version 3.0.6). Py-Pi, https://pypi.org/project/sounderpy/
+	     Gillett, K., 2025: SounderPy: Vertical Profile Data Retrieval & Analysis Tool for Python (Version 3.0.8). Py-Pi, https://pypi.org/project/sounderpy/
 
 
 ***************************************************************
@@ -163,8 +164,6 @@ Thanks for checking out and using SounderPy. My name is Kyle Gillett and I'm a m
 
 SounderPy is published on PyPi and the source code is available on GitHub -- this tool is an open source project. If you have found SounderPy useful in your work, I'd love to hear about it! The coolest part of this project has been hearing how many folks have been using this software. If you'd like to support continued SounderPy development, consider "`Buying me a coffee <https://www.buymeacoffee.com/kylejgillett>`_"! ☕. 
 
-*Have an issue?* You can `open a GitHub Issue <https://github.com/kylejgillett/sounderpy/issues>`_ or just shoot me a DM on `Twitter <https://twitter.com/wxkylegillett>`_!
-
 
 ****************************************
 
@@ -183,3 +182,38 @@ Useful Links
 
 Thanks for using SounderPy!
 
+
+
+****************************************
+
+
+Publications Using SounderPy
+-----------------------------
+
+- Barton, B., and C. Gormley, 2025: Analysis of strongly tornadic environments in Central and Eastern Europe utilizing ERA5 reanalysis data. M.S. thesis, University of Oklahoma
+
+- Capuli, G., M. A. Noveno, and M. P. Ibañez, 2025: A case study of the tornadic supercell in the Province of Pampanga, Philippines (27 May 2024). arXiv, https://doi.org/10.48550/arXiv.2504.20559
+
+- Capuli, G. H., 2024: Friday the 13th hailstorm in the province of Bulacan, Philippines (13 August 2021): A case study. arXiv, https://arxiv.org/abs/2412.09307
+
+- Capuli, G., 2024: Project Severe Weather Archive of the Philippines (SWAP) Part 1: Establishing a baseline climatology for severe weather across the Philippine Archipelago. Ann. Geophys., 67(5), GC553, https://doi.org/10.4401/ag-9151
+
+- Coffer, B. E., M. D. Parker, M. C. Coniglio, and C. R. Homeyer, 2025: Supercell environments using GridRad-Severe and the HRRR: Addressing discrepancies between prior tornado datasets. Wea. Forecasting, 40, 1405–1428, https://doi.org/10.1175/WAF-D-24-0251.1
+
+- Hua, Z., A. Anderson-Frey, M. C. Brown, and Q. Jiang, 2025: A data-driven explainable framework for diagnosing cluster assignments of right-moving tornadic supercell soundings. Preprint, 4 Jul 2025.
+
[…]
