### 2025-08-22 bb38d7f
edit LICENSES wording
