+++ joss-paper/paper.md
@@ -82,7 +82,7 @@ Scott Thomas (National Weather Service), Ethan O’Neill (University of North
 Carolina at Charlotte), Brian Rakoczy (Ohio State University), Daryl Herzmann (Iowa State
 University), Amelia R.H. Urquhart (University of Oklahoma), Ryan Vandersmith, and many others.
 
-This project also extends its graditude to the researchers, institutions, and 
+This project also extends its gratitude to the researchers, institutions, and 
 individuals who have utilized SounderPy in their work and publications, driving
 its growth and application within the meteorological community.
 
