### 2025-08-22 402a915
edit LICENSES wording
