+++ joss-paper/paper.md
@@ -16,6 +16,7 @@ affiliations:
     index: 1
 date: 3 March 2025
 bibliography: paper.bib
+
 ---
 
 
