### 2025-08-22 #8087 @kylejgillett
Hi @pdebuyl.

Thank you very much! Those changes have been pushed and the docs/paper are being generated now.

Thanks very much for your insight and help!

> Hello [@kylejgillett](https://github.com/kylejgillett) thanks for the update. Regarding the license, can you specify "SounderPy contains the source code to SHARPpy (Sounding/Hodograph Analysis and Research Program in Python) in src/sounderpy/SHARPPYMAIN." ? This will specify to which files that separate license applies, which is important. Can you specify somewhere in the doc (in troubleshooting for instance) the fixed reference image? (for example: "To check that sounderpy produces the expected result, run the following (...) and compare to the file `images/test-figure_abr-2024-08-29-00z.png`").
>
> Typo in the manuscript: "graditude" -> "gratitude".
>
> Once the above is done, the paper is good to go :-)
