### 2025-08-22 #8087 @kylejgillett
The new docs & updated paper.pdf are finished
