### 2025-08-22 #8087 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
